+++ src/index.ts
@@ -16,6 +16,7 @@ type LoggerOptions = {
     level: LogLevel;
     prefix: string;
     timestamp: boolean;
+    callerInfo: boolean;
 };
 
 /**
@@ -25,11 +26,15 @@ type LoggerOptions = {
 export function getLogger(options?: Partial<LoggerOptions>) {
     const defaultSettings: LoggerOptions = {
         level: _resolveLogLevel(options?.level, process.env.LOG_LEVEL as keyof typeof LogLevel),
-        prefix: '',
-        timestamp: false,
+        prefix: process.env.LOG_PREFIX ?? '',
+        timestamp: process.env.LOG_TIMESTAMP === 'true' || false,
+        callerInfo: process.env.LOG_CALLERINFO === 'true' || false,
     };
 
     const settings: LoggerOptions = Object.assign(defaultSettings, options || {});
+
+    const masks = new Map();
+    let hasMasks = false;
 
     return {
         trace: (message: string, attachment?: object): void => {
@@ -59,25 +64,61 @@ export function getLogger(options?: Partial<LoggerOptions>) {
         getSettings: (): LoggerOptions => {
             return settings;
         },
+        isDebugEnabled: (): boolean => {
+            return settings.level <= LogLevel.DEBUG;
+        },
+        isTraceEnabled: (): boolean => {
+            return settings.level <= LogLevel.TRACE;
+        },
+        addMask(value: string, placeholder = '***'): void {
+            masks.set(value, placeholder);
+            hasMasks = true;
+        },
+        removeMask(value: string): void {
+            masks.delete(value);
+            hasMasks = masks.size > 0;
+        },
+        clearMasks(): void {
+            masks.clear();
+            hasMasks = false;
+        },
     };
 
-    function _log(logLevel: LogLevel, message: string, attachment?: object): void {
+    function _log(logLevel: LogLevel, message: string, attachment?: Record<string, any>): void {
         if (settings.level <= logLevel) {
             let formattedMessage;
-
             if (settings.timestamp) {
                 formattedMessage = format('[%s] %s %s%s', LogLevel[logLevel], new Date().toISOString(), settings.prefix, message);
             } else {
                 formattedMessage = format('[%s] %s%s', LogLevel[logLevel], settings.prefix, message);
             }
 
+            if (settings.callerInfo) {
+                attachment = _attachCallerInfo(attachment);
+            }
+
+            if (hasMasks) {
+                formattedMessage = _mask(formattedMessage);
+            }
+
             /* tslint:disable: no-console */
             if (attachment) {
-                console.log(formattedMessage, inspect(attachment, false, null));
+                let stringAttachment = inspect(attachment, false, null);
+                if (hasMasks) {
+                    stringAttachment = _mask(stringAttachment);
+                }
+                console.log(formattedMessage, stringAttachment);
             } else {
                 console.log(formattedMessage);
             }
             /* tslint:enable: no-console */
+        }
+
+        function _mask(input: string): string {
+            for (const [maskValue, maskPlaceholder] of masks.entries()) {
+                input = input.replaceAll(maskValue, maskPlaceholder);
+            }
+            return input;
         }
     }
 }
@@ -99,4 +140,24 @@ function _resolveLogLevel(level?: LogLevel, processLogLevel?: keyof typeof LogLe
     }
 
     return LogLevel.INFO;
+}
+
+/* istanbul ignore next */
+function _attachCallerInfo(attachment?: Record<string, any>): Record<string, any> {
+    if (attachment == null) {
+        attachment = {};
+    }
+    const callerStackLine = new Error().stack?.split('\n')[4];
+    if (callerStackLine) {
+        const pattern = /\((.*):(\d+):(\d+)\)$/; // This regex pattern matches the file path, line number, and column number
+        const match = pattern.exec(callerStackLine);
+
+        if (match) {
+            attachment.callerInfo = {
+                file: match[1],
+                line: match[2],
+            };
+        }
+    }
+    return attachment;
 }